requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Video processing
opencv-python==4.9.0.80
Pillow==10.2.0

# AI/ML models
sentence-transformers==2.3.1
google-generativeai==0.3.2
torch==2.1.2

# Data handling
numpy==1.26.3
pandas==2.2.0

# Utilities
tqdm==4.66.1
requests==2.31.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Optional: OpenAI alternative
# openai==1.10.0
